Fix errors, final commit (for now)
In this commit, I revised the querySelectors for the first three
variables, to remove a functional error preventing the rest of the
application from working properly. I may return to edit the CSS on this
to make it look nice, and perhaps add connections for finding local
grocery stores, coupons for list items, or other fun additions. Until
then, this is the final commit for this project.